requirements.txt
# Core AI/Agent Dependencies
anthropic>=0.40.0
langgraph>=0.2.50

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0

# Google Sheets Integration
gspread>=6.0.0
google-auth>=2.23.0
oauth2client>=4.1.3

# Email Services
sendgrid>=6.11.0

# Configuration
python-dotenv>=1.0.0